+++ packages/ice/src/service/ServerRunner.ts
@@ -108,12 +108,13 @@ class ServerRunner extends Runner {
       },
     }, 'esbuild', { isServer: true });
     const { alias, ignores } = filterAlias(task.config.alias || {});
+    const define = getRuntimeDefination(task.config.define || {});
     const runtimeMeta = new RuntimeMeta({
       rootDir,
       alias,
       ignores,
       external: server.externals || [],
-      define: getRuntimeDefination(task.config.define || {}),
+      define,
     });
 
     const esbuildPlugins = [
@@ -272,12 +273,12 @@ class ServerRunner extends Runner {
             const bundlePath = await runtimeMeta.bundle(id);
             return { externalize: bundlePath };
           }
-
           return {
             code: await transformJsxRuntime(code),
           };
         }
       },
+      define,
     });
   }