+++ packages/ice/src/service/serverCompiler.ts
@@ -64,9 +64,12 @@ export const getRuntimeDefination = (
   const runtimeDefineVars = {
     ...runtimeVars,
   };
-  // auto stringify define value
+  // esbuild only receive the string type of replacement expressions, so we need to transform the boolean to string.
+  // link: https://esbuild.github.io/api/#define
   Object.keys(defineVars).forEach((key) => {
-    stringifiedDefine[key] = JSON.stringify(defineVars[key]);
+    stringifiedDefine[key] = typeof defineVars[key] === 'string'
+      ? defineVars[key]
+      : JSON.stringify(defineVars[key]);
   });
   // get runtime variable for server build
   Object.keys(process.env).forEach((key) => {