+++ packages/miniapp-runtime/src/app/index.ts
@@ -3,6 +3,7 @@ import runClientApp from './runClientApp.js';
 import Link from './Link.js';
 import useSearchParams from './useSearchParams.js';
 import { routerHistory as history } from './history.js';
+import usePageLifecycle from './usePageLifecycle.js';
 import { useData, useConfig } from './routeContext.js';
 
 export {
@@ -16,4 +17,5 @@ export {
   useSearchParams,
   history,
   defineDataLoader,
+  usePageLifecycle,
 };