+++ packages/runtime/src/index.ts
@@ -16,7 +16,7 @@ import type {
   RouteWrapper,
   RenderMode,
   DistType,
-  DataLoaderConfig,
+  Loader,
   RouteWrapperConfig,
 } from './types.js';
 import Runtime from './runtime.js';
@@ -49,8 +49,9 @@ import { routerHistory as history } from './history.js';
 import KeepAliveOutlet from './KeepAliveOutlet.js';
 import ClientOnly from './ClientOnly.js';
 import useMounted from './useMounted.js';
+import usePageLifecycle from './usePageLifecycle.js';
 import { withSuspense, useSuspenseData } from './Suspense.js';
-import { createRouteLoader, WrapRouteComponent, RouteErrorComponent } from './routes.js';
+import { createRouteLoader, WrapRouteComponent, RouteErrorComponent, Await } from './routes.js';
 
 export {
   getAppConfig,
@@ -87,10 +88,14 @@ export {
   KeepAliveOutlet,
   AppErrorBoundary,
   ClientOnly,
+
   useMounted,
+  usePageLifecycle,
 
   withSuspense,
   useSuspenseData,
+
+  Await,
 
   createRouteLoader,
   WrapRouteComponent,
@@ -109,7 +114,7 @@ export type {
   RouteWrapper,
   RenderMode,
   DistType,
-  DataLoaderConfig,
+  Loader,
   RunClientAppOptions,
   MetaType,
   TitleType,